Refactor Fusegu API documentation and enhance health check endpoint
- Updated OpenAPI documentation to include new transaction analysis endpoints and improved health check descriptions.
- Refactored health response model import for clarity.
- Adjusted validation ranges for session age and quantity in request models to prevent overflow.
- Enhanced feature store interface with additional methods for transaction and user analysis.
- Updated transaction service to utilize a mutex for rule engine access, ensuring thread safety during evaluations.
- Added a new method for generating secure random strings in the crypto utility module.

+++ src/models/email.rs
@@ -52,7 +52,7 @@ impl EmailAddress {
 
     /// Calculate email reputation score
     pub fn reputation_score(&self) -> f64 {
-        let mut score: f32 = 80.0; // Start with good reputation
+        let mut score: f64 = 80.0; // Start with good reputation
         
         if self.is_high_risk {
             score -= 40.0;
